feat/freeBoard

+++ client/src/components/board/freeBoard/postContent.tsx
@@ -40,17 +40,21 @@ function PostContent() {
   let elapsedTime: number = CalElapsedTime(listData.createdAt);
   let calTime = '';
 
-  if (elapsedTime < 60) {
-    calTime = '방금 전';
-  } else if (elapsedTime < 3600) {
-    elapsedTime = Math.round(elapsedTime / 60);
-    calTime = `${elapsedTime}분 전`;
-  } else if (elapsedTime < 43200) {
-    elapsedTime = Math.round(elapsedTime / 3600);
-    calTime = `${elapsedTime}시간 전`;
-  } else {
-    elapsedTime = Math.round(elapsedTime / 43200);
-    calTime = `${elapsedTime}일 전`;
+  if (!isPending) {
+    if (elapsedTime < 60) {
+      calTime = '방금 전';
+    } else if (elapsedTime < 3600) {
+      elapsedTime = Math.round(elapsedTime / 60);
+      calTime = `${elapsedTime}분 전`;
+    } else if (elapsedTime < 43200) {
+      elapsedTime = Math.round(elapsedTime / 3600);
+      calTime = `${elapsedTime}시간 전`;
+    } else if (elapsedTime < 129600) {
+      elapsedTime = Math.round(elapsedTime / 43200);
+      calTime = `${elapsedTime}일 전`;
+    } else {
+      calTime = listData.createdAt.slice(0, 24);
+    }
   }
 
   const fetchPostDetail = async () => {
@@ -71,16 +75,17 @@ function PostContent() {
 
   const fetchDeletePost = async () => {
     try {
-      if (urlData === '/free') {
-        // 삭제 확인 경고문 넣어야함
-        await DeletePost('frees', idData);
-        alert('게시물을 삭제하였습니다.');
-        navigate('/free');
-      } else {
-        // 삭제 확인 경고문 넣어야함
-        await DeletePost('qnas/answers', idData);
-        alert('게시물을 삭제하였습니다.');
-        navigate('/qna');
+      const confirm = window.confirm('게시글을 삭제하시겠습니까?');
+      if (confirm) {
+        if (urlData === '/free') {
+          await DeletePost('frees', idData);
+          alert('게시물을 삭제하였습니다.');
+          navigate('/free');
+        } else {
+          await DeletePost('qnas/answers', idData);
+          alert('게시물을 삭제하였습니다.');
+          navigate('/qna');
+        }
       }
     } catch (err) {
       console.error(err);
